Add Support for "Use Name"/"Use Title" Flag
Based on a feature request. If a calendar event has "Use Name" or "Use Title" in the _description_ of the event, then the in/out board will show the name instead of the location. This is useful if, for example, the location of the event is "Joe's Urology, 1234 Medical Lane" but you don't want that to be displayed on the board. Enter the event's title as "Doctors Appointment" and description as "Use Title" and the board will display that instead.